Move copying/cloning in extra methods of Gmp
Passing the copy parameter already doesn't work as expected. It seems
that the other parameters send over the wire are ignored by gvmd or even
like with task or target name lead to an error. Therefore introduce
clone methods that just expect an id as parameter and remove the copy
parameter from all create methods.