Implement a generalized timeout for reading a xml response
The timeout is a minimum time before aborting. It could be possible e.g.
that reading the socket will take 59 seconds for reading a chunk. If the
timeout is 60 seconds the procedure will be aborted after >= 118
seconds only.

But IMHO reducing the timeout to 5 seconds for reading a chunk,
catching the OSError timeout exception and failing more early is
necessary here.