Adjust handling of 'asset_type' in create_asset
It turns out the `asset_type` element is _not_ mandatory for the
`create_asset` command. However, if it is present, the value will be
ignored by the server and replaced with the default `host` asset type.

This commit accounts for these circumstances and will raise an
`InvalidArgument` exception to prevent clients from sending values which
will silently be ignored on the server side.