Access `Mapping` via `collections.abc`
The Collections Abstract Base Classes were moved from the `collections`
module to the `collections.abc` module in Python 3.3. Using them via the
`collections` module was deprecated in Python 3.7 and will stop working
in Python 3.8, so the `isinstance` calls in `python-gvm` should be
adjusted accordingly.

For reference, this the warning shown when running the code in Python
3.7:
```
DeprecationWarning: Using or importing the ABCs from 'collections'
instead of from 'collections.abc' is deprecated, and in 3.8 it will stop
working
```