Adapt modify_tag to implementation
Adapt the parameter validation in `modify_tag` to the actual
implementation in `gmvd` and adjust tests and documentation accordingly.

The previous implementation refused a number of combinations which are
are accepted by the `gvmd` from GVM-10. Most notably, the
`resource_type` is rarely required, in contrast to the `create_tag`
command. Also, sending *only* the `resource_type` is acceptable for
modifying the type of resource a tag is intended for.