Change modify_user
Modify user now only takes an ID for identifying
the user which is to be modified.
The name parameter is now the new name and not an
identifier anymore.
The comment, auth_source and group_ids parameters
got added. Now its possible to change those fields
too.
auth_source has to be used explicitly if not the
default Source allowed for authentication for
the user is to be used.