Cleanup project (#1083)
* Remove: Delete unused and obsolete changelog files

* Remove: Drop pontos changelog generation config

Just use the defaults of pontos.

* Remove: Drop the poetry.toml file

Let the user decide where to store the virtual environment for
python-gvm. For some setups (for example in a CI) it might be better if
the virtual environment is stored outside of the project.

* Update copyright in README

* Change: Drop support for Python 3.7 and 3.8 and add Python 3.12

We are supporting Python 3.9 - 3.12 now only.

* Update dependencies for Python changes

Also make the dependencies more loose.